requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
seaborn>=0.11.0
openai>=1.0.0
anthropic>=0.3.0
google-generativeai>=0.3.0  # Add this for Gemini support
transformers>=4.30.0
torch>=2.0.0
python-dotenv>=0.19.0
